# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Students app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of AquaTrack app"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
...
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user information",
        "operationId": "getUserById",
        "description": "Retrieve detailed information about a user by their ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique identifier of the user.",
              "example": "63f12345e4567ab89cde9012"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user information.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully retrieved user information."
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "_id",
                        "name",
                        "weight",
                        "activeTime",
                        "email",
                        "waterDailyNorma",
                        "gender",
                        "avatar",
                        "createdAt",
                        "updatedAt"
                      ],
                      "properties": {
                        "_id": {
                          "description": "User's id",
                          "type": "string",
                          "example": "65e4decdd286b30065d54af9"
                        },
                        "name": {
                          "description": "User's name",
                          "type": "string",
                          "example": "John Doe"
                        },
                        "weight": {
                          "description": "User's weight",
                          "type": "number",
                          "example": 64
                        },
                        "activeTime": {
                          "description": "User's active time",
                          "type": "number",
                          "example": 2
                        },
                        "email": {
                          "description": "User's email",
                          "type": "string",
                          "example": "joedoe@gmail.com"
                        },
                        "waterDailyNorma": {
                          "description": "The amount of water a user should drink",
                          "type": "number",
                          "example": 1.4
                        },
                        "gender": {
                          "description": "User's gender",
                          "type": "string",
                          "enum": ["male", "female"]
                        },
                        "avatar": {
                          "description": "User's avatar",
                          "type": "string",
                          "example": "https://res.cloudinary.com/uqwgdu/image/upload/image.png"
                        },
                        "createdAt": {
                          "description": "The date and time when the user was created.",
                          "type": "string",
                          "example": "2024-12-29T20:33:45Z"
                        },
                        "updatedAt": {
                          "description": "The date and time when the user information was last updated.",
                          "type": "string",
                          "example": "2024-12-30T10:15:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/paths/~1water/get/responses/401"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not found"
                    },
                    "data": {
                      "type": "object",
                      "required": ["message"],
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "User with id 67719abf3fc799eedfd041ba not found`"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update user record by id",
        "operationId": "updateUser",
        "description": "Update uers record by providing the id and new data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            },
            "description": "The unique identifier of the user record to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Update user's name",
                    "example": "Bob Sahe"
                  },
                  "weight": {
                    "type": "number",
                    "description": "Update user's weight",
                    "example": 54
                  },
                  "activeTime": {
                    "type": "number",
                    "description": "Update user's active time",
                    "example": 1
                  },
                  "waterDailyNorma": {
                    "type": "number",
                    "description": "Update user's water daily norma",
                    "example": 1.1
                  },
                  "gender": {
                    "type": "string",
                    "description": "Update user's gender",
                    "example": "male"
                  },
                  "avatar": {
                    "type": "string",
                    "description": "Update user's avatar",
                    "example": "https://res.cloudinary.com/uqwgdu/image/upload/image.png"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated user record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated user record with id 65ca67e7ae7f10c88b598384!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/paths/~1users~1%7Bid%7D/get/responses/200/content/application~1json/schema/properties/data"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1water/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1water/get/responses/401"
          },
          "404": {
            "$ref": "#/paths/~1users~1%7Bid%7D/get/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}